Rename /abi to /vote, and add AddressBasedAuthorizer class to /vote as discussed. Fix test.
