+++ rskj-core/src/main/java/co/rsk/config/BridgeConstants.java
@@ -21,7 +21,7 @@ package co.rsk.config;
 import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.Coin;
 import co.rsk.bitcoinj.core.NetworkParameters;
-import co.rsk.peg.AddressBasedAuthorizer;
+import co.rsk.peg.vote.AddressBasedAuthorizer;
 import co.rsk.peg.federation.Federation;
 import java.util.List;
