+++ rskj-core/src/main/java/co/rsk/peg/BridgeSupport.java
@@ -58,9 +58,9 @@ import co.rsk.config.BridgeConstants;
 import co.rsk.core.RskAddress;
 import co.rsk.crypto.Keccak256;
 import co.rsk.panic.PanicProcessor;
-import co.rsk.peg.abi.ABICallElection;
-import co.rsk.peg.abi.ABICallSpec;
-import co.rsk.peg.abi.ABICallVoteResult;
+import co.rsk.peg.vote.ABICallElection;
+import co.rsk.peg.vote.ABICallSpec;
+import co.rsk.peg.vote.ABICallVoteResult;
 import co.rsk.peg.bitcoin.BitcoinUtils;
 import co.rsk.peg.bitcoin.CoinbaseInformation;
 import co.rsk.peg.bitcoin.MerkleBranch;
@@ -80,6 +80,7 @@ import co.rsk.peg.utils.BtcTransactionFormatUtils;
 import co.rsk.peg.utils.PartialMerkleTreeFormatUtils;
 import co.rsk.peg.utils.RejectedPegoutReason;
 import co.rsk.peg.utils.UnrefundablePeginReason;
+import co.rsk.peg.vote.AddressBasedAuthorizer;
 import co.rsk.peg.whitelist.LockWhitelist;
 import co.rsk.peg.whitelist.LockWhitelistEntry;
 import co.rsk.peg.whitelist.OneOffWhiteListEntry;
