+++ rskj-core/src/main/java/co/rsk/peg/vote/ABICallElection.java
@@ -16,10 +16,9 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package co.rsk.peg.abi;
+package co.rsk.peg.vote;
 
 import co.rsk.core.RskAddress;
-import co.rsk.peg.AddressBasedAuthorizer;
 
 import java.util.*;
