+++ rskj-core/src/test/java/co/rsk/peg/BridgeTestIntegration.java
@@ -50,7 +50,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.BiFunction;
 
-import co.rsk.peg.abi.ABICallSpec;
+import co.rsk.peg.vote.ABICallSpec;
 import co.rsk.peg.federation.Federation;
 import co.rsk.peg.federation.FederationMember;
 import co.rsk.test.builders.BlockChainBuilder;
